+++ TFT_eSPIDisplay.h
@@ -42,11 +42,11 @@ public:
   /// @param text Text to be displayed on this row
   /// @param underlined (Optional) Flag to underline this row - default false
   /// @param column (Optional) Column to start displaying the text, column being width of a character (not pixels)
-  void displayRow(int row, const char *text, bool underlined = false, int column = 0) override;
+  void displayRow(uint8_t row, const char *text, bool underlined = false, uint8_t column = 0) override;
 
   /// @brief Clear the specified row
   /// @param row Row number as specified in the SCREEN() command (not pixels)
-  void clearRow(int row) override;
+  void clearRow(uint8_t row) override;
 
   /// @brief Display the startup screen with software version
   /// @param version EX-Display version
