+++ test/mocks/MockDisplay.h
@@ -28,9 +28,9 @@ public:
 
   MOCK_METHOD(void, clearScreen, (), (override));
 
-  MOCK_METHOD(void, displayRow, (int row, const char *text, bool underlined, int column), (override));
+  MOCK_METHOD(void, displayRow, (uint8_t row, const char *text, bool underlined, uint8_t column), (override));
 
-  MOCK_METHOD(void, clearRow, (int row), (override));
+  MOCK_METHOD(void, clearRow, (uint8_t row), (override));
 
   MOCK_METHOD(void, displayStartupInfo, (const char *version), (override));
 };
